+++ src/servers/sapphire_zone/Script/NativeScriptApi.h
@@ -88,7 +88,7 @@ public:
       ScriptObject( questId, typeid( EventScript ).hash_code() )
    { }
 
-   virtual void onTalk( uint32_t eventId, Entity::Player& player, uint64_t actorId ) { }
+   virtual void onTalk( uint32_t eventId, Entity::Player& player, uint32_t actorId ) { }
    virtual void onNpcKill( uint32_t npcId, Entity::Player& player ) { }
    virtual void onEmote( uint64_t actorId, uint32_t eventId, uint32_t emoteId, Entity::Player& player ) { }
    virtual void onEnterTerritory( Entity::Player& player, uint32_t eventId, uint16_t param1, uint16_t param2 ) { }
